@@ -1,20 +1,20 @@
 name: bitcoin_flutter
 description: A dart Bitcoin library for Flutter. BIP32, BIP39, P2PKH integration.
-version: 2.0.2
-homepage: https://github.com/anicdh
-author: andoan <anic.dh@gmail.com>
+version: 1.0.0
+homepage: https://github.com/ThorbenD/bitcoin_flutter
+author: Tommy <dreier.thorben@gmail.com>
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
-  bip39: ^1.0.3
-  bip32: ^1.0.5
-  pointycastle: ^1.0.2
-  hex: ^0.1.2
-  bs58check: ^1.0.1
-  meta: ^1.2.3
-  bech32: 0.1.2
+  bip39: ^1.0.6
+  bip32: ^2.0.0
+  pointycastle: ^3.5.2
+  hex: ^0.2.0
+  bs58check: ^1.0.2
+  meta: ^1.7.0
+  bech32: ^0.2.1
 
 dev_dependencies:
-  test: ^1.15.4
+  test: ^1.20.1
