@@ -5,39 +5,40 @@ import 'package:bip32/src/utils/wif.dart' as wif;
 import 'models/networks.dart';
 
 class ECPair {
-  Uint8List _d;
-  Uint8List _Q;
+  Uint8List? _d;
+  Uint8List? _Q;
   NetworkType network;
   bool compressed;
-  ECPair(Uint8List _d, Uint8List _Q, {network, compressed}) {
-    this._d = _d;
-    this._Q = _Q;
-    this.network = network ?? bitcoin;
-    this.compressed = compressed ?? true;
-  }
-  Uint8List get publicKey {
-    if (_Q == null) _Q = ecc.pointFromScalar(_d, compressed);
+
+  ECPair(Uint8List? this._d, Uint8List? this._Q,
+      {NetworkType? network, bool? compressed})
+      : network = network ?? bitcoin,
+        compressed = compressed ?? true {}
+
+  Uint8List? get publicKey {
+    if (_Q == null) _Q = ecc.pointFromScalar(_d!, compressed);
     return _Q;
   }
 
-  Uint8List get privateKey => _d;
+  Uint8List? get privateKey => _d;
+
   String toWIF() {
     if (privateKey == null) {
       throw new ArgumentError('Missing private key');
     }
     return wif.encode(new wif.WIF(
-        version: network.wif, privateKey: privateKey, compressed: compressed));
+        version: network.wif, privateKey: privateKey!, compressed: compressed));
   }
 
   Uint8List sign(Uint8List hash) {
-    return ecc.sign(hash, privateKey);
+    return ecc.sign(hash, privateKey!);
   }
 
   bool verify(Uint8List hash, Uint8List signature) {
-    return ecc.verify(hash, publicKey, signature);
+    return ecc.verify(hash, publicKey!, signature);
   }
 
-  factory ECPair.fromWIF(String w, {NetworkType network}) {
+  factory ECPair.fromWIF(String w, {NetworkType? network}) {
     wif.WIF decoded = wif.decode(w);
     final version = decoded.version;
     // TODO support multi networks
@@ -57,16 +58,18 @@ class ECPair {
     return ECPair.fromPrivateKey(decoded.privateKey,
         compressed: decoded.compressed, network: nw);
   }
+
   factory ECPair.fromPublicKey(Uint8List publicKey,
-      {NetworkType network, bool compressed}) {
+      {NetworkType? network, bool? compressed}) {
     if (!ecc.isPoint(publicKey)) {
       throw new ArgumentError('Point is not on the curve');
     }
     return new ECPair(null, publicKey,
         network: network, compressed: compressed);
   }
+
   factory ECPair.fromPrivateKey(Uint8List privateKey,
-      {NetworkType network, bool compressed}) {
+      {NetworkType? network, bool? compressed}) {
     if (privateKey.length != 32)
       throw new ArgumentError(
           'Expected property privateKey of type Buffer(Length: 32)');
@@ -75,8 +78,9 @@ class ECPair {
     return new ECPair(privateKey, null,
         network: network, compressed: compressed);
   }
+
   factory ECPair.makeRandom(
-      {NetworkType network, bool compressed, Function rng}) {
+      {NetworkType? network, bool? compressed, Function? rng}) {
     final rfunc = rng ?? _randomBytes;
     Uint8List d;
 //    int beginTime = DateTime.now().millisecondsSinceEpoch;
@@ -89,7 +93,8 @@ class ECPair {
   }
 }
 
-const int _SIZE_BYTE = 255;
+const int _SIZE_BYTE = 256;
+
 Uint8List _randomBytes(int size) {
   final rng = Random.secure();
   final bytes = Uint8List(size);
