@@ -1,24 +1,24 @@
-import 'package:meta/meta.dart';
-
 class NetworkType {
   String messagePrefix;
-  String bech32;
+  String? bech32;
   Bip32Type bip32;
   int pubKeyHash;
   int scriptHash;
   int wif;
+  int opreturnSize;
 
-  NetworkType(
-      {@required this.messagePrefix,
-      this.bech32,
-      @required this.bip32,
-      @required this.pubKeyHash,
-      @required this.scriptHash,
-      @required this.wif});
+  NetworkType({required this.messagePrefix,
+    this.bech32,
+    required this.bip32,
+    required this.pubKeyHash,
+    required this.scriptHash,
+    required this.wif,
+    required this.opreturnSize});
 
   @override
   String toString() {
-    return 'NetworkType{messagePrefix: $messagePrefix, bech32: $bech32, bip32: ${bip32.toString()}, pubKeyHash: $pubKeyHash, scriptHash: $scriptHash, wif: $wif}';
+    return 'NetworkType{messagePrefix: $messagePrefix, bech32: $bech32, bip32: ${bip32.toString()}, '
+        'pubKeyHash: $pubKeyHash, scriptHash: $scriptHash, wif: $wif op return max size: $opreturnSize}';
   }
 }
 
@@ -26,7 +26,7 @@ class Bip32Type {
   int public;
   int private;
 
-  Bip32Type({@required this.public, @required this.private});
+  Bip32Type({required this.public, required this.private});
 
   @override
   String toString() {
@@ -40,7 +40,9 @@ final bitcoin = new NetworkType(
     bip32: new Bip32Type(public: 0x0488b21e, private: 0x0488ade4),
     pubKeyHash: 0x00,
     scriptHash: 0x05,
-    wif: 0x80);
+    wif: 0x80,
+    opreturnSize: 80
+);
 
 final testnet = new NetworkType(
     messagePrefix: '\x18Bitcoin Signed Message:\n',
@@ -48,4 +50,6 @@ final testnet = new NetworkType(
     bip32: new Bip32Type(public: 0x043587cf, private: 0x04358394),
     pubKeyHash: 0x6f,
     scriptHash: 0xc4,
-    wif: 0xef);
+    wif: 0xef,
+    opreturnSize: 80
+);
