@@ -1,7 +1,6 @@
 import 'dart:typed_data';
-import 'package:meta/meta.dart';
-import 'package:bip32/src/utils/ecurve.dart' show isPoint;
 import 'package:bech32/bech32.dart';
+import 'package:bip32/src/utils/ecurve.dart' show isPoint;
 
 import '../crypto.dart';
 import '../models/networks.dart';
@@ -14,9 +13,9 @@ class P2WPKH {
 
   PaymentData data;
   NetworkType network;
-  P2WPKH({@required data, network}) {
-    this.network = network ?? bitcoin;
-    this.data = data;
+
+  P2WPKH({required this.data, NetworkType? network})
+      : network = network ?? bitcoin {
     _init();
   }
 
@@ -28,51 +27,52 @@ class P2WPKH {
         data.witness == null) throw new ArgumentError('Not enough data');
 
     if (data.address != null) {
-      _getDataFromAddress(data.address);
+      _getDataFromAddress(data.address!);
     }
 
     if (data.hash != null) {
       _getDataFromHash();
     }
 
-    if (data.output != null) {
-      if (data.output.length != 22 ||
-          data.output[0] != OPS['OP_0'] ||
-          data.output[1] != 20) // 0x14
+    final output = data.output;
+    if (output != null) {
+      if (output.length != 22 ||
+          output[0] != OPS['OP_0'] ||
+          output[1] != 20) // 0x14
         throw new ArgumentError('Output is invalid');
       if (data.hash == null) {
-        data.hash = data.output.sublist(2);
+        data.hash = output.sublist(2);
       }
       _getDataFromHash();
     }
 
     if (data.pubkey != null) {
-      data.hash = hash160(data.pubkey);
+      data.hash = hash160(data.pubkey!);
       _getDataFromHash();
     }
 
-    if (data.witness != null) {
-      if (data.witness.length != 2)
-        throw new ArgumentError('Witness is invalid');
-      if (!bscript.isCanonicalScriptSignature(data.witness[0]))
+    final witness = data.witness;
+    if (witness != null) {
+      if (witness.length != 2) throw new ArgumentError('Witness is invalid');
+      if (!bscript.isCanonicalScriptSignature(witness[0]))
         throw new ArgumentError('Witness has invalid signature');
-      if (!isPoint(data.witness[1]))
+      if (!isPoint(witness[1]))
         throw new ArgumentError('Witness has invalid pubkey');
-      _getDataFromWitness(data.witness);
+      _getDataFromWitness(witness);
     } else if (data.pubkey != null && data.signature != null) {
-      data.witness = [data.signature, data.pubkey];
+      data.witness = [data.signature!, data.pubkey!];
       if (data.input == null) data.input = EMPTY_SCRIPT;
     }
   }
 
-  void _getDataFromWitness([List<Uint8List> witness]) {
+  void _getDataFromWitness(List<Uint8List> witness) {
     if (data.input == null) {
       data.input = EMPTY_SCRIPT;
     }
     if (data.pubkey == null) {
       data.pubkey = witness[1];
       if (data.hash == null) {
-        data.hash = hash160(data.pubkey);
+        data.hash = hash160(data.pubkey!);
       }
       _getDataFromHash();
     }
@@ -81,7 +81,7 @@ class P2WPKH {
 
   void _getDataFromHash() {
     if (data.address == null) {
-      data.address = segwit.encode(Segwit(network.bech32, 0, data.hash));
+      data.address = segwit.encode(Segwit(network.bech32!, 0, data.hash!));
     }
     if (data.output == null) {
       data.output = bscript.compile([OPS['OP_0'], data.hash]);
