@@ -1,3 +1,4 @@
+import 'package:bitcoin_flutter/src/utils/constants/op.dart';
 import 'package:test/test.dart';
 import 'dart:io';
 import 'dart:convert';
@@ -21,17 +22,13 @@ constructSign(f, TransactionBuilder txb) {
     if (inputs[i]['signs'] == null) continue;
     (inputs[i]['signs'] as List<dynamic>).forEach((sign) {
       ECPair keyPair = ECPair.fromWIF(sign['keyPair'], network: network);
-      txb.sign(
-          vin: i,
-          keyPair: keyPair,
-          witnessValue: sign['value'],
-          hashType: sign['hashType']);
+      txb.sign(vin: i, keyPair: keyPair, witnessValue: sign['value'], hashType: sign['hashType']);
     });
   }
   return txb;
 }
 
-TransactionBuilder construct(f, [bool dontSign]) {
+TransactionBuilder construct(f, [bool? dontSign]) {
   final network = NETWORKS[f['network']];
   final txb = new TransactionBuilder(network: network);
   if (f['version'] != null) txb.setVersion(f['version']);
@@ -66,26 +63,20 @@ TransactionBuilder construct(f, [bool dontSign]) {
 }
 
 main() {
-  final fixtures = json.decode(
-      new File('test/fixtures/transaction_builder.json')
-          .readAsStringSync(encoding: utf8));
+  final fixtures = json.decode(new File('test/fixtures/transaction_builder.json').readAsStringSync(encoding: utf8));
   group('TransactionBuilder', () {
-    final keyPair = ECPair.fromPrivateKey(HEX.decode(
-        '0000000000000000000000000000000000000000000000000000000000000001'));
-    final scripts = [
-      '1BgGZ9tcN4rm9KBzDn7KprQz87SZ26SAMH',
-      '1cMh228HTCiwS8ZsaakH8A8wze1JR5ZsP'
-    ].map((x) => Address.addressToOutputScript(x));
-    final txHash = HEX.decode(
-        '0e7cea811c0be9f73c0aca591034396e7264473fc25c1ca45195d7417b36cbe2');
+    final keyPair = ECPair.fromPrivateKey(
+        HEX.decode('0000000000000000000000000000000000000000000000000000000000000001') as Uint8List);
+    final scripts = ['1BgGZ9tcN4rm9KBzDn7KprQz87SZ26SAMH', '1cMh228HTCiwS8ZsaakH8A8wze1JR5ZsP']
+        .map((x) => Address.addressToOutputScript(x));
+    final txHash = HEX.decode('0e7cea811c0be9f73c0aca591034396e7264473fc25c1ca45195d7417b36cbe2');
     group('fromTransaction', () {
       (fixtures['valid']['build'] as List<dynamic>).forEach((f) {
         test('returns TransactionBuilder, with ${f['description']}', () {
           final network = NETWORKS[f['network'] ?? 'bitcoin'];
           final tx = Transaction.fromHex(f['txHex']);
           final txb = TransactionBuilder.fromTransaction(tx, network);
-          final txAfter =
-              f['incomplete'] != null ? txb.buildIncomplete() : txb.build();
+          final txAfter = f['incomplete'] != null ? txb.buildIncomplete() : txb.build();
           expect(txAfter.toHex(), f['txHex']);
           expect(txb.network, network);
         });
@@ -95,10 +86,8 @@ main() {
           final tx = new Transaction();
           f['inputs'] as List<dynamic>
             ..forEach((input) {
-              final txHash2 = Uint8List.fromList(
-                  HEX.decode(input['txId']).reversed.toList());
-              tx.addInput(txHash2, input['vout'], null,
-                  bscript.fromASM(input['scriptSig']));
+              final txHash2 = Uint8List.fromList(HEX.decode(input['txId']).reversed.toList());
+              tx.addInput(txHash2, input['vout'], null, bscript.fromASM(input['scriptSig']));
             });
           f['outputs'] as List<dynamic>
             ..forEach((output) {
@@ -109,12 +98,10 @@ main() {
           final txAfter = f['incomplete'] ? txb.buildIncomplete() : txb.build();
 
           for (var i = 0; i < txAfter.ins.length; i++) {
-            test(bscript.toASM(txAfter.ins[i].script),
-                f['inputs'][i]['scriptSigAfter']);
+            test(bscript.toASM(txAfter.ins[i].script!), f['inputs'][i]['scriptSigAfter']);
           }
           for (var i = 0; i < txAfter.outs.length; i++) {
-            test(bscript.toASM(txAfter.outs[i].script),
-                f['outputs'][i]['script']);
+            test(bscript.toASM(txAfter.outs[i].script!), f['outputs'][i]['script']);
           }
         });
       });
@@ -131,7 +118,7 @@ main() {
         });
     });
     group('addInput', () {
-      TransactionBuilder txb;
+      late TransactionBuilder txb;
       setUp(() {
         txb = new TransactionBuilder();
       });
@@ -171,30 +158,24 @@ main() {
         expect(txb.addInput(txHash, 0), 0);
         expect(txb.addInput(txHash, 1), 1);
       });
-      test(
-          'throws if SIGHASH_ALL has been used to sign any existing scriptSigs',
-          () {
+      test('throws if SIGHASH_ALL has been used to sign any existing scriptSigs', () {
         txb.addInput(txHash, 0);
         txb.addOutput(scripts.elementAt(0), 1000);
         txb.sign(vin: 0, keyPair: keyPair);
         try {
           expect(txb.addInput(txHash, 0), isArgumentError);
         } catch (err) {
-          expect((err as ArgumentError).message,
-              'No, this would invalidate signatures');
+          expect((err as ArgumentError).message, 'No, this would invalidate signatures');
         }
       });
     });
     group('addOutput', () {
-      TransactionBuilder txb;
+      late TransactionBuilder txb;
       setUp(() {
         txb = new TransactionBuilder();
       });
       test('accepts an address string and value', () {
-        final address =
-            new P2PKH(data: new PaymentData(pubkey: keyPair.publicKey))
-                .data
-                .address;
+        final address = new P2PKH(data: new PaymentData(pubkey: keyPair.publicKey)).data.address;
         final vout = txb.addOutput(address, 1000);
         expect(vout, 0);
         final txout = txb.tx.outs[0];
@@ -210,11 +191,9 @@ main() {
       });
       test('throws if address is of the wrong network', () {
         try {
-          expect(txb.addOutput('2NGHjvjw83pcVFgMcA7QvSMh2c246rxLVz9', 1000),
-              isArgumentError);
+          expect(txb.addOutput('2NGHjvjw83pcVFgMcA7QvSMh2c246rxLVz9', 1000), isArgumentError);
         } catch (err) {
-          expect((err as ArgumentError).message,
-              'Invalid version or Network mismatch');
+          expect((err as ArgumentError).message, 'Invalid version or Network mismatch');
         }
       });
       test('add second output after signed first input with SIGHASH_NONE', () {
@@ -228,8 +207,7 @@ main() {
         txb.sign(vin: 0, keyPair: keyPair, hashType: SIGHASH_NONE);
         expect(txb.addOutput(scripts.elementAt(0), 2000), 0);
       });
-      test('add second output after signed first input with SIGHASH_SINGLE',
-          () {
+      test('add second output after signed first input with SIGHASH_SINGLE', () {
         txb.addInput(txHash, 0);
         txb.addOutput(scripts.elementAt(0), 2000);
         txb.sign(vin: 0, keyPair: keyPair, hashType: SIGHASH_SINGLE);
@@ -241,21 +219,93 @@ main() {
         try {
           expect(txb.addOutput(scripts.elementAt(0), 2000), isArgumentError);
         } catch (err) {
-          expect((err as ArgumentError).message,
-              'No, this would invalidate signatures');
+          expect((err as ArgumentError).message, 'No, this would invalidate signatures');
         }
       });
-      test(
-          'throws if SIGHASH_ALL has been used to sign any existing scriptSigs',
-          () {
+      test('throws if SIGHASH_ALL has been used to sign any existing scriptSigs', () {
         txb.addInput(txHash, 0);
         txb.addOutput(scripts.elementAt(0), 2000);
         txb.sign(vin: 0, keyPair: keyPair);
         try {
           expect(txb.addOutput(scripts.elementAt(1), 9000), isArgumentError);
         } catch (err) {
-          expect((err as ArgumentError).message,
-              'No, this would invalidate signatures');
+          expect((err as ArgumentError).message, 'No, this would invalidate signatures');
+        }
+      });
+    });
+    group('addNullOutput', () {
+      late TransactionBuilder txb;
+      late String data;
+      late String data2;
+
+      setUp(() {
+        txb = new TransactionBuilder();
+        data = 'Hey this is a random string without coins. Extended to 80 characters............';
+        data2 = 'And this is another string.';
+      });
+
+      expectOutputScript(input, Uint8List expectPushData) {
+        final opReturn = OPS['OP_RETURN']!;
+        final expectScript = bscript.compile([opReturn, expectPushData]);
+        final vout = txb.addNullOutput(input);
+        expect(vout, 0);
+        final txout = txb.tx.outs[0];
+        expect(txout.script, expectScript);
+        expect(txout.value, 0);
+      }
+
+      test('accepts a string', () {
+        final rawData = Uint8List.fromList(utf8.encode(data));
+        expectOutputScript(data, rawData);
+      });
+
+      test('accepts Uint8List data', () {
+        final rawData = Uint8List.fromList(List.generate(80, (i) => i));
+        expectOutputScript(rawData, rawData);
+      });
+
+      test('throws if too much data is provided', () {
+        try {
+          expect(txb.addNullOutput('Lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi sagittis placerat.'),
+              isArgumentError);
+        } catch (err) {
+          expect((err as ArgumentError).message, 'Too much data, max OP_RETURN size is 80');
+        }
+      });
+      test('add second output after signed first input with SIGHASH_NONE', () {
+        txb.addInput(txHash, 0);
+        txb.addNullOutput(data);
+        txb.sign(vin: 0, keyPair: keyPair, hashType: SIGHASH_NONE);
+        expect(txb.addNullOutput(data2), 1);
+      });
+      test('add first output after signed first input with SIGHASH_NONE', () {
+        txb.addInput(txHash, 0);
+        txb.sign(vin: 0, keyPair: keyPair, hashType: SIGHASH_NONE);
+        expect(txb.addNullOutput(data), 0);
+      });
+      test('add second output after signed first input with SIGHASH_SINGLE', () {
+        txb.addInput(txHash, 0);
+        txb.addNullOutput(data);
+        txb.sign(vin: 0, keyPair: keyPair, hashType: SIGHASH_SINGLE);
+        expect(txb.addNullOutput(data2), 1);
+      });
+      test('add first output after signed first input with SIGHASH_SINGLE', () {
+        txb.addInput(txHash, 0);
+        txb.sign(vin: 0, keyPair: keyPair, hashType: SIGHASH_SINGLE);
+        try {
+          expect(txb.addNullOutput(data), isArgumentError);
+        } catch (err) {
+          expect((err as ArgumentError).message, 'No, this would invalidate signatures');
+        }
+      });
+      test('throws if SIGHASH_ALL has been used to sign any existing scriptSigs', () {
+        txb.addInput(txHash, 0);
+        txb.addNullOutput(data);
+        txb.sign(vin: 0, keyPair: keyPair);
+        try {
+          expect(txb.addNullOutput(data2), isArgumentError);
+        } catch (err) {
+          expect((err as ArgumentError).message, 'No, this would invalidate signatures');
         }
       });
     });
@@ -268,42 +318,31 @@ main() {
         try {
           expect(txb.setLockTime(65535), isArgumentError);
         } catch (err) {
-          expect((err as ArgumentError).message,
-              'No, this would invalidate signatures');
+          expect((err as ArgumentError).message, 'No, this would invalidate signatures');
         }
       });
     });
     group('sign', () {
       fixtures['invalid']['sign'] as List<dynamic>
         ..forEach((f) {
-          test(
-              'throws ${f['exception']} ${f['description'] != null ? f['description'] : ''}',
-              () {
+          test('throws ${f['exception']} ${f['description'] != null ? f['description'] : ''}', () {
             final txb = construct(f, true);
             var threw = false;
             final inputs = f['inputs'] as List;
             for (var i = 0; i < inputs.length; i++) {
               inputs[i]['signs'] as List<dynamic>
                 ..forEach((sign) {
-                  final keyPairNetwork =
-                      NETWORKS[sign['network'] ?? f['network']];
-                  final keyPair2 =
-                      ECPair.fromWIF(sign['keyPair'], network: keyPairNetwork);
+                  final keyPairNetwork = NETWORKS[sign['network'] ?? f['network']];
+                  final keyPair2 = ECPair.fromWIF(sign['keyPair'], network: keyPairNetwork);
                   if (sign['throws'] != null && sign['throws']) {
                     try {
-                      expect(
-                          txb.sign(
-                              vin: i,
-                              keyPair: keyPair2,
-                              hashType: sign['hashType']),
-                          isArgumentError);
+                      expect(txb.sign(vin: i, keyPair: keyPair2, hashType: sign['hashType']), isArgumentError);
                     } catch (err) {
                       expect((err as ArgumentError).message, f['exception']);
                     }
                     threw = true;
                   } else {
-                    txb.sign(
-                        vin: i, keyPair: keyPair2, hashType: sign['hashType']);
+                    txb.sign(vin: i, keyPair: keyPair2, hashType: sign['hashType']);
                   }
                 });
             }
@@ -316,8 +355,7 @@ main() {
         ..forEach((f) {
           test('builds ${f['description']}', () {
             final txb = construct(f);
-            final tx =
-                f['incomplete'] != null ? txb.buildIncomplete() : txb.build();
+            final tx = f['incomplete'] != null ? txb.buildIncomplete() : txb.build();
 
             expect(tx.toHex(), f['txHex']);
           });
@@ -329,8 +367,7 @@ main() {
               try {
                 TransactionBuilder txb;
                 if (f['txHex'] != null) {
-                  txb = TransactionBuilder.fromTransaction(
-                      Transaction.fromHex(f['txHex']));
+                  txb = TransactionBuilder.fromTransaction(Transaction.fromHex(f['txHex']));
                 } else {
                   txb = construct(f);
                 }
@@ -345,8 +382,7 @@ main() {
                 try {
                   TransactionBuilder txb;
                   if (f['txHex'] != null) {
-                    txb = TransactionBuilder.fromTransaction(
-                        Transaction.fromHex(f['txHex']));
+                    txb = TransactionBuilder.fromTransaction(Transaction.fromHex(f['txHex']));
                   } else {
                     txb = construct(f);
                   }
@@ -359,8 +395,7 @@ main() {
               test('does not throw if buildIncomplete', () {
                 TransactionBuilder txb;
                 if (f['txHex'] != null) {
-                  txb = TransactionBuilder.fromTransaction(
-                      Transaction.fromHex(f['txHex']));
+                  txb = TransactionBuilder.fromTransaction(Transaction.fromHex(f['txHex']));
                 } else {
                   txb = construct(f);
                 }
