@@ -1,6 +1,3 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library bitcoin_flutter;
 
 export 'src/bitcoin_flutter_base.dart';
@@ -12,4 +9,5 @@ export 'src/ecpair.dart';
 export 'src/payments/p2pkh.dart';
 export 'src/payments/p2wpkh.dart';
 export 'src/payments/index.dart';
+
 // TODO: Export any libraries intended for clients of this package.
