@@ -16,7 +16,7 @@ const SCRIPT_TYPES = {
   'WITNESS_COMMITMENT': 'witnesscommitment'
 };
 
-String classifyOutput(Uint8List script) {
+String? classifyOutput(Uint8List script) {
   if (witnessPubKeyHash.outputCheck(script)) return SCRIPT_TYPES['P2WPKH'];
   if (pubkeyhash.outputCheck(script)) return SCRIPT_TYPES['P2PKH'];
   final chunks = bscript.decompile(script);
@@ -24,7 +24,7 @@ String classifyOutput(Uint8List script) {
   return SCRIPT_TYPES['NONSTANDARD'];
 }
 
-String classifyInput(Uint8List script) {
+String? classifyInput(Uint8List script) {
   final chunks = bscript.decompile(script);
   if (chunks == null) throw new ArgumentError('Invalid script');
   if (pubkeyhash.inputCheck(chunks)) return SCRIPT_TYPES['P2PKH'];
@@ -32,7 +32,7 @@ String classifyInput(Uint8List script) {
   return SCRIPT_TYPES['NONSTANDARD'];
 }
 
-String classifyWitness(List<Uint8List> script) {
+String? classifyWitness(List<Uint8List> script) {
   final chunks = bscript.decompile(script);
   if (chunks == null) throw new ArgumentError('Invalid script');
   if (witnessPubKeyHash.inputCheck(chunks)) return SCRIPT_TYPES['P2WPKH'];
