@@ -4,9 +4,9 @@ import '../../src/crypto.dart';
 import 'varuint.dart';
 import '../../src/models/networks.dart';
 
-Uint8List magicHash(String message, [NetworkType network]) {
+Uint8List magicHash(String message, [NetworkType? network]) {
   network = network ?? bitcoin;
-  Uint8List messagePrefix = utf8.encode(network.messagePrefix);
+  Uint8List messagePrefix = utf8.encode(network.messagePrefix) as Uint8List;
   int messageVISize = encodingLength(message.length);
   int length = messagePrefix.length + messageVISize + message.length;
   Uint8List buffer = new Uint8List(length);
