@@ -4,8 +4,7 @@ import 'package:bip39/bip39.dart' as bip39;
 
 void main() {
   group('bitcoin-dart (HDWallet)', () {
-    var seed = bip39.mnemonicToSeed(
-        'praise you muffin lion enable neck grocery crumble super myself license ghost');
+    var seed = bip39.mnemonicToSeed('praise you muffin lion enable neck grocery crumble super myself license ghost');
     HDWallet hdWallet = new HDWallet.fromSeed(seed);
     test('valid seed', () {
       expect(hdWallet.seed,
