@@ -1,7 +1,9 @@
+import 'dart:typed_data';
+
 import '../../lib/src/models/networks.dart' as NETWORKS;
 import '../../lib/src/ecpair.dart' show ECPair;
 import '../../lib/src/payments/index.dart' show PaymentData;
-import '../../lib/src/payments/p2pkh.dart' show P2PKH, P2PKHData;
+import '../../lib/src/payments/p2pkh.dart' show P2PKH;
 import '../../lib/src/payments/p2wpkh.dart' show P2WPKH;
 import 'package:pointycastle/digests/sha256.dart';
 import 'dart:convert';
@@ -12,7 +14,8 @@ NETWORKS.NetworkType litecoin = new NETWORKS.NetworkType(
     bip32: new NETWORKS.Bip32Type(public: 0x019da462, private: 0x019d9cfe),
     pubKeyHash: 0x30,
     scriptHash: 0x32,
-    wif: 0xb0);
+    wif: 0xb0,
+    opreturnSize: 80);
 // deterministic RNG for testing only
 rng(int number) {
   return utf8.encode('zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz');
@@ -30,7 +33,7 @@ main() {
     });
     test('can generate an address from a SHA256 hash', () {
       final hash = new SHA256Digest()
-          .process(utf8.encode('correct horse battery staple'));
+          .process(utf8.encode('correct horse battery staple') as Uint8List);
       final keyPair = ECPair.fromPrivateKey(hash);
       final address =
           new P2PKH(data: new PaymentData(pubkey: keyPair.publicKey))
@@ -52,8 +55,8 @@ main() {
       final keyPair = ECPair.makeRandom(network: testnet, rng: rng);
       final wif = keyPair.toWIF();
       final address = new P2PKH(
-              data: new PaymentData(pubkey: keyPair.publicKey),
-              network: testnet)
+          data: new PaymentData(pubkey: keyPair.publicKey),
+          network: testnet)
           .data
           .address;
       expect(address, 'mubSzQNtZfDj1YdNP6pNDuZy6zs6GDn61L');
@@ -63,8 +66,8 @@ main() {
       final keyPair = ECPair.makeRandom(network: litecoin, rng: rng);
       final wif = keyPair.toWIF();
       final address = new P2PKH(
-              data: new PaymentData(pubkey: keyPair.publicKey),
-              network: litecoin)
+          data: new PaymentData(pubkey: keyPair.publicKey),
+          network: litecoin)
           .data
           .address;
       expect(address, 'LZJSxZbjqJ2XVEquqfqHg1RQTDdfST5PTn');
@@ -84,8 +87,8 @@ main() {
       final keyPair = ECPair.fromWIF(
           'cPaJYBMDLjQp5gSUHnBfhX4Rgj95ekBS6oBttwQLw3qfsKKcDfuB');
       final address = new P2WPKH(
-              data: new PaymentData(pubkey: keyPair.publicKey),
-              network: testnet)
+          data: new PaymentData(pubkey: keyPair.publicKey),
+          network: testnet)
           .data
           .address;
       expect(address, 'tb1qgmp0h7lvexdxx9y05pmdukx09xcteu9sx2h4ya');
