@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:bitcoin_flutter/src/models/networks.dart';
 import 'package:bitcoin_flutter/src/payments/index.dart' show PaymentData;
 import 'package:bitcoin_flutter/src/payments/p2pkh.dart';
@@ -43,7 +45,8 @@ void main() {
     test('can create a BIP32, bitcoin, account 0, external address', () {
       const path = "m/0'/0/0";
       final root = bip32.BIP32.fromSeed(HEX.decode(
-          'dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd'));
+          'dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd')
+      as Uint8List);
       final child1 = root.derivePath(path);
       // option 2, manually
       final child1b = root.deriveHardened(0).derive(0).derive(0);
@@ -52,7 +55,8 @@ void main() {
     });
     test('can create a BIP44, bitcoin, account 0, external address', () {
       final root = bip32.BIP32.fromSeed(HEX.decode(
-          'dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd'));
+          'dddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddd')
+      as Uint8List);
       final child1 = root.derivePath("m/44'/0'/0'/0/0");
       // option 2, manually
       final child1b = root
@@ -88,7 +92,7 @@ void main() {
 }
 
 String getAddress(node, [network]) {
-  return P2PKH(data: new PaymentData(pubkey: node.publicKey), network: network)
+  return P2PKH(data: PaymentData(pubkey: node.publicKey), network: network)
       .data
-      .address;
+      .address!;
 }
